+++ packages/scripts/config/webpack.config.merge.js
@@ -1,23 +1,24 @@
 const mergeWith = require('lodash.mergeWith');
-const defaultConfig = require('./webpack.config');
-
+const { getAbsolutePath, getUserConfig } = require('../scripts/utils');
 
 // Default configuration file
-const configurationFiles = [defaultConfig];
+const webpackConfigurations = [require('./webpack.config')];
 
 // Dev configuration file
-if (process.env.TALEND_MODE === 'development') {
-	configurationFiles.push(require('./webpack.config.dev'));
+const mode = process.env.TALEND_MODE || 'production';
+if (mode === 'development') {
+	webpackConfigurations.push(require('./webpack.config.dev'));
 }
 
 // App configuration file
-const appConfigPath = process.env.TALEND_APP_CONFIG;
-if (appConfigPath) {
-	console.log(`Merge ${process.env.TALEND_MODE} webpack config with app custom one (${appConfigPath})`);
-	configurationFiles.push(require(appConfigPath));
+const userConfigPath = getUserConfig(['webpack', 'config', mode]);
+if (userConfigPath) {
+	const userConfigAbsolutePath = getAbsolutePath(userConfigPath);
+	console.log(`Merge ${mode} webpack config with custom one (${userConfigAbsolutePath})`);
+	webpackConfigurations.push(require(userConfigAbsolutePath));
 }
 
-module.exports = mergeWith({}, ...configurationFiles, function customizer(objValue, srcValue) {
+module.exports = mergeWith({}, ...webpackConfigurations, function customizer(objValue, srcValue) {
 	if (Array.isArray(objValue)) {
 		return objValue.concat(srcValue);
 	}
