+++ packages/scripts/package.json
@@ -25,6 +25,7 @@
     "extract-text-webpack-plugin": "3.0.2",
     "file-loader": "1.1.5",
     "html-webpack-plugin": "2.28.0",
+    "lodash.get": "^4.4.2",
     "lodash.mergewith": "^4.6.1",
     "node-sass": "4.7.2",
     "postcss-loader": "2.0.8",
