+++ packages/scripts/scripts/start.js
@@ -1,70 +1,17 @@
 #!/usr/bin/env node
 const spawn = require('cross-spawn');
-const yargsParser = require('yargs-parser');
-const { getAbsolutePath, resolveBin, hereRelative } = require('./utils');
+const { getEnv, hereRelative, printLogo, resolveBin } = require('./utils');
 
 const webpackDevServer = resolveBin('webpack-dev-server');
 
-const args = process.argv.slice(2);
-const parsedArgs = yargsParser(args);
-const env = Object.create(process.env);
-
-console.log('##########################################################################################################################################################################################################\n' +
-	' */  */                   &**/                                                                               %#                                                                                  # \n' +
-	'/      *                  ****                                                                            #####                                                                               ###### \n' +
-	'       #                      (                                                                           #####                                                                               ###### \n' +
-	'/(    /( *(                    /&                              ###                                      #####                                                                               ###### \n' +
-	'             **                  *                            (#####                                      #####                                                                               ###### \n' +
-	'                  *(               (                          (#####                                      #####                                                                               ###### \n' +
-	'                      */            *                         (#####                   &%##%%&            #####                %%#%%                         &%%%%                     &%%%%%##### \n' +
-	'                          **          *    &((&               (#############     %################        #####            #############&        #####   ############%            %################# \n' +
-	'                              (*(      (**********(           (############      ###################      #####          #################       ######################         #################### \n' +
-	'                                   */ **************          (#####              #/          #######     #####        #####       ######&     #########       #######      ########        ###### \n' +
-	'       /**(                          ****************         (#####                           ######     #####        #####          ######     ######           ######     #######          ###### \n' +
-	'      ***********************************************         (#####                           ######     #####       ######           ######    ######           ######    #######           ###### \n' +
-	'       ****                          ****************         (#####               %#################     #####       #######################    ######           ######    ######            ###### \n' +
-	'                                    #***************(         (#####             ####################     #####       #######################    ######           ######    ######            ###### \n' +
-	'                             (  %*(    ************           (#####           #######         ######     #####       ######                     ######           ######    ######            ###### \n' +
-	'                           (***(       *  /*****              (#####           ######          ######     #####       ######                     ######           ######    ######            ###### \n' +
-	'                             (       #/                       (#####&          ######          ######     #####        ######&                   ######           ######     ######           ###### \n' +
-	'                                    *                          ######     @    #######       ########     ######        #######&         ##      ######           ######     /#######       ######## \n' +
-	'                                   (                           #############    #####################     ##########     ###################     ######           ######       #####################&\n' +
-	'                                 *                              #############     ###########(  #####%     ###########      ###############      ######           ######         ############  ######\n' +
-	'                           ****(/                                                                                                                                                                    \n' +
-	'                         *      /                                                                                                                                                                    \n' +
-	'                         *       *                                                                                                                                                                   \n' +
-	'                         ((     *                                                                                                                                                                    \n' +
-	'\n' +
-	'                          .----------------.  .----------------.  .----------------.  .----------------.  .----------------.  .----------------.  .----------------. \n' +
-	'                         | .--------------. || .--------------. || .--------------. || .--------------. || .--------------. || .--------------. || .--------------. |\n' +
-	'                         | |    _______   | || |     ______   | || |  _______     | || |     _____    | || |   ______     | || |  _________   | || |    _______   | |\n' +
-	'                         | |   /  ___  |  | || |   .\' ___  |  | || | |_   __ \\    | || |    |_   _|   | || |  |_   __ \\   | || | |  _   _  |  | || |   /  ___  |  | |\n' +
-	'                         | |  |  (__ \\_|  | || |  / .\'   \\_|  | || |   | |__) |   | || |      | |     | || |    | |__) |  | || | |_/ | | \\_|  | || |  |  (__ \\_|  | |\n' +
-	'                         | |   \'.___`-.   | || |  | |         | || |   |  __ /    | || |      | |     | || |    |  ___/   | || |     | |      | || |   \'.___`-.   | |\n' +
-	'                         | |  |`\\____) |  | || |  \\ `.___.\'\\  | || |  _| |  \\ \\_  | || |     _| |_    | || |   _| |_      | || |    _| |_     | || |  |`\\____) |  | |\n' +
-	'                         | |  |_______.\'  | || |   `._____.\'  | || | |____| |___| | || |    |_____|   | || |  |_____|     | || |   |_____|    | || |  |_______.\'  | |\n' +
-	'                         | |              | || |              | || |              | || |              | || |              | || |              | || |              | |\n' +
-	'                         | \'--------------\' || \'--------------\' || \'--------------\' || \'--------------\' || \'--------------\' || \'--------------\' || \'--------------\' |\n' +
-	'                          \'----------------\'  \'----------------\'  \'----------------\'  \'----------------\'  \'----------------\'  \'----------------\'  \'----------------\' \n' +
-	'##########################################################################################################################################################################################################');
+printLogo();
 
 console.log('\nCONFIGURATION ------------------------------------------------------------------------------------\n');
 
-// USER : Pass extra config to merge with common config.
-// talend-scripts-start --config ./webpack.config.js
-const appConfigPath = parsedArgs.config;
-if (appConfigPath) {
-	const appConfigResolvedPath = getAbsolutePath(appConfigPath);
-	console.log(`App extra webpack config : ${appConfigResolvedPath}`);
-	env.TALEND_APP_CONFIG = appConfigResolvedPath;
-}
-
-// USER : Pass api url to proxy /api routes
-// talend-scripts-start --api-url http://localhost:8080
-const apiUrl = parsedArgs['api-url'];
-if (apiUrl) {
-	console.log(`Proxy /api to : ${apiUrl}`);
-	env.TALEND_API_URL = apiUrl;
+// USER : current env vars and talend scripts configuration in <project-folder>/talend-scripts.json
+const env = getEnv();
+if (env.TALEND_SCRIPTS_CONFIG) {
+	console.log('Talend scripts configuration file found and loaded');
 }
 
 // INTERNAL : Set the mode to get devServer config
@@ -83,7 +30,3 @@ const result = spawn.sync(
 	{ stdio: 'inherit', env });
 
 process.exit(result.status);
-
-// TODO waiting for
-// - resolve-url-loader : https://github.com/bholloway/resolve-url-loader/issues/79
-// - extract-text-webpack-plugin : https://github.com/webpack-contrib/extract-text-webpack-plugin/issues/701
