+++ packages/scripts/scripts/utils.js
@@ -1,4 +1,5 @@
 const fs = require('fs');
+const get = require('lodash.get');
 const path = require('path');
 const which = require('which');
 
@@ -40,8 +41,70 @@ function hereRelative(dirname, p) {
 		.replace(process.cwd(), '.');
 }
 
+function getEnv() {
+	const env = Object.create(process.env);
+
+	const userConfigFilePath = path.join(process.cwd(), 'talend-scripts.json');
+	if (fs.existsSync(userConfigFilePath)) {
+		env.TALEND_SCRIPTS_CONFIG = require(userConfigFilePath);
+	}
+
+	return env;
+}
+
+function getUserConfig(configObjectPath) {
+	return get(
+		process.env.TALEND_SCRIPTS_CONFIG,
+		configObjectPath
+	);
+}
+
+function printLogo() {
+	console.log('##########################################################################################################################################################################################################\n' +
+		' */  */                   &**/                                                                               %#                                                                                  # \n' +
+		'/      *                  ****                                                                            #####                                                                               ###### \n' +
+		'       #                      (                                                                           #####                                                                               ###### \n' +
+		'/(    /( *(                    /&                              ###                                      #####                                                                               ###### \n' +
+		'             **                  *                            (#####                                      #####                                                                               ###### \n' +
+		'                  *(               (                          (#####                                      #####                                                                               ###### \n' +
+		'                      */            *                         (#####                   &%##%%&            #####                %%#%%                         &%%%%                     &%%%%%##### \n' +
+		'                          **          *    &((&               (#############     %################        #####            #############&        #####   ############%            %################# \n' +
+		'                              (*(      (**********(           (############      ###################      #####          #################       ######################         #################### \n' +
+		'                                   */ **************          (#####              #/          #######     #####        #####       ######&     #########       #######      ########        ###### \n' +
+		'       /**(                          ****************         (#####                           ######     #####        #####          ######     ######           ######     #######          ###### \n' +
+		'      ***********************************************         (#####                           ######     #####       ######           ######    ######           ######    #######           ###### \n' +
+		'       ****                          ****************         (#####               %#################     #####       #######################    ######           ######    ######            ###### \n' +
+		'                                    #***************(         (#####             ####################     #####       #######################    ######           ######    ######            ###### \n' +
+		'                             (  %*(    ************           (#####           #######         ######     #####       ######                     ######           ######    ######            ###### \n' +
+		'                           (***(       *  /*****              (#####           ######          ######     #####       ######                     ######           ######    ######            ###### \n' +
+		'                             (       #/                       (#####&          ######          ######     #####        ######&                   ######           ######     ######           ###### \n' +
+		'                                    *                          ######     @    #######       ########     ######        #######&         ##      ######           ######     /#######       ######## \n' +
+		'                                   (                           #############    #####################     ##########     ###################     ######           ######       #####################&\n' +
+		'                                 *                              #############     ###########(  #####%     ###########      ###############      ######           ######         ############  ######\n' +
+		'                           ****(/                                                                                                                                                                    \n' +
+		'                         *      /                                                                                                                                                                    \n' +
+		'                         *       *                                                                                                                                                                   \n' +
+		'                         ((     *                                                                                                                                                                    \n' +
+		'\n' +
+		'                          .----------------.  .----------------.  .----------------.  .----------------.  .----------------.  .----------------.  .----------------. \n' +
+		'                         | .--------------. || .--------------. || .--------------. || .--------------. || .--------------. || .--------------. || .--------------. |\n' +
+		'                         | |    _______   | || |     ______   | || |  _______     | || |     _____    | || |   ______     | || |  _________   | || |    _______   | |\n' +
+		'                         | |   /  ___  |  | || |   .\' ___  |  | || | |_   __ \\    | || |    |_   _|   | || |  |_   __ \\   | || | |  _   _  |  | || |   /  ___  |  | |\n' +
+		'                         | |  |  (__ \\_|  | || |  / .\'   \\_|  | || |   | |__) |   | || |      | |     | || |    | |__) |  | || | |_/ | | \\_|  | || |  |  (__ \\_|  | |\n' +
+		'                         | |   \'.___`-.   | || |  | |         | || |   |  __ /    | || |      | |     | || |    |  ___/   | || |     | |      | || |   \'.___`-.   | |\n' +
+		'                         | |  |`\\____) |  | || |  \\ `.___.\'\\  | || |  _| |  \\ \\_  | || |     _| |_    | || |   _| |_      | || |    _| |_     | || |  |`\\____) |  | |\n' +
+		'                         | |  |_______.\'  | || |   `._____.\'  | || | |____| |___| | || |    |_____|   | || |  |_____|     | || |   |_____|    | || |  |_______.\'  | |\n' +
+		'                         | |              | || |              | || |              | || |              | || |              | || |              | || |              | |\n' +
+		'                         | \'--------------\' || \'--------------\' || \'--------------\' || \'--------------\' || \'--------------\' || \'--------------\' || \'--------------\' |\n' +
+		'                          \'----------------\'  \'----------------\'  \'----------------\'  \'----------------\'  \'----------------\'  \'----------------\'  \'----------------\' \n' +
+		'##########################################################################################################################################################################################################');
+}
+
 module.exports = {
-	resolveBin,
 	getAbsolutePath,
+	getEnv,
+	getUserConfig,
 	hereRelative,
+	printLogo,
+	resolveBin,
 };
